fix(VSelectionControlGroup): provided props & defaults target
